sensor done
